[⚙]🐛fix: 회원 탈퇴 데이터 반영

+++ src/main/java/ddwu/project/mdm_ver2/domain/product/dto/ProductResponse.java
@@ -22,6 +22,6 @@ public class ProductResponse {
     private String imgUrl;
     private Character favState;
     private List<Review> reviewList;
-    private int reviewCnt;
+    private Long reviewCnt;
     private float reviewStarAvg;
 }
